requirements.txt
mmif-python>=1.0.14
pandas>=2
